requirements.txt
arrow==1.1.1
asgiref==3.4.1
blessed==1.18.1
certifi==2021.5.30
charset-normalizer==2.0.5
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.7.0
django-filter==2.4.0
django-picklefield==3.0.1
django-q==1.3.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
model-bakery==1.3.2
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
whitenoise==5.3.0
